requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
shapely==2.0.2
geopandas==0.14.1
requests>=2.26.0
python-dotenv==1.0.0
python-multipart==0.0.6
aiofiles>=0.7.0
tqdm>=4.65.0
psutil>=5.9.0
prometheus-client>=0.17.0
psycopg[binary]>=3.1.12
GeoAlchemy2==0.14.2
sqlalchemy-utils>=0.41.1
pandas>=2.0.0
numpy==1.24.3
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
pre-commit>=3.3.0
flask==2.3.3
rasterio==1.3.8
pillow==10.0.0
gunicorn==21.2.0
